New experimental scheduler (currently disabled).
The old scheduler is ``optimal'' in the sense that it minimizes
register pressure.  The only way to reduce register pressure is to
schedule dependent instructions as closely as possible, so as to
minimize the life time of registers.  This strategy maximizes the
number of pipeline stalls, however.  With enough registers and short
enough pipelines, this tradeoff is fine.  This is no longer the case
for the devilish pipeline of the Pentium IV or (probably) the PowerPC
970.

The new scheduler switches to a ``list scheduler'' for dags smaller
than a specified size.  The list scheduler executes a butterfly left
to right one column at the time.  This amounts to the best possible
pipeline utilization, and the worst possible register pressure.

The ``specified size'' defaults to 0, i.e., no change from fftw2 and
fftw-3.0.  It seems like a value of 7--10 produces the best results
for Pentium IV (probably screwing the G3/G4 powerpcs and sparc, but I
haven't tried.)  As time goes by, we may want to increase this number
to favor newer processors over older processors.