Fixed conditions under which the rank-geq2-rdft2 solver is applicable.
The old solver was not applicable for out-of-place problems
unless DESTROY_INPUT.  This is bogus.  As long as the subsolvers
honor !DESTROY_INPUT, the solver is always applicable.

Changed semantics of test program, so that PRESERVE_INPUT is always
true unless the problem specifies destroy_input explicitly.  Without
this change, there is no way to test the new solver.