added unsolvable check for rdft2 problem
An in-place rdft2 problem is ill-formed if the real data, including the
extra "padding" elements, do not coincide with the complex data.

CHANGE: the new code considers all in-place split r2c and c2r problems
to be ill-formed.  Previously, these could be done, but only if the
entire multi-dimensional array fit into the buffer, which is kind of
stupid.  I'm not sure it's worth it to even try to support the
split in-place r2c case.