do not assume that a semaphore can be freed just because nobody is using it
Let S be a semaphore, initially 0.  Let thread A execute UP(S);
let thread B execute DOWN(S); free(&S);  It is unclear whether this
code is correct with posix semaphores.  The problem is whether UP()
uses S after allowing DOWN() to continue; this seems to be the
case in the glibc-2.7 implementation, and thus the pattern above
seems to be incorrect.  Avoid using such a pattern, and introduce
a global semaphore for the unavoidable case when nothing else
can be depended upon.