Add r2cb_2.c
r2cb_2.c is needed for problem rb2, which is not equivalent to
rf2 (unlike kb2, which is equivalent to kf2).

This change would not matter much except that rb2 is generated
when reducing backward rdft2 to dft, and the absence of the codelet was
preventing radix 2 from being employed at all in this case.