Add VZEROUPPER at the end of AVX codelets
If the Intel Optimization Manual is to be believed, we need to wave a
dead chicken before transitioning from AVX code to SSE code.  I am
supposed to believe that there is a transition penalty for doing so,
unless one uses a magic VZEROUPPER instruction that apparently has
zero cost.  Whatever.