use malloc() instead of alloca() for large buffers
The proximate cause for this patch is that OpenBSD/i386 reserves 256KB
stack size per thread.  We were allocating a buffer of size
128*130*sizeof(fftw_complex) that exceeds the stack.

While 128*130*sizeof(fftw_complex) = 260KiB is the worst case for
normal configurations, it is a good idea to limit stack allocation
just in case.  Also, the generic solver might in principle generate
unbounded buffers, even though it is normally disabled for n > 137.

So, as an added precaution, we now never stack-allocate buffers larger
than 64KiB, which ought to be enough for anybody.