make the index-computation logic less paranoid
The problem is that for each K and for each expression of the form P[I
+ STRIDE * K] in a loop, most compilers will try to lift an induction
variable PK := &P[I + STRIDE * K].  In large codelets we have many
such values of K.  For example, a codelet of size 32 with 4 input
pointers will generate O(128) induction variables, which will likely
overflow the register set, which is likely worse than doing the index
computation in the first place.

In the past we (wisely and correctly) assumed that compilers will do
the wrong thing, and consequently we disabled the induction-variable
"optimization" altogether by setting STRIDE ^= ZERO, where ZERO is a
value guaranteed to be 0.  Since the compiler does not know that
ZERO=0, it cannot perform its "optimization" and it is forced to
behave sensibly.

With this patch, FFTW is a little bit less paranoid.  FFTW now
disables the induction-variable optimization" only when we estimate
that the codelet uses more than ESTIMATED_AVAILABLE_INDEX_REGISTERS
induction variables.

Currently we set ESTIMATED_AVAILABLE_INDEX_REGISTERS=16.  16 registers ought
to be enough for anybody (or so the amd64 and ARM ISA's seem to imply).