Define the version number as M4 macros, so that it is defined only once.
We used to have two version strings: the package number FFTW-X.Y.Z and
the libtool number CURRENT:REVISION:AGE with the invariant that
REVISION==Z.  Unfortunately in the fftw-3.3.3 release we forgot to
change REVISION, with the result that fftw-3.3.3 and fftw-3.3.2 have
the same REVISION number.

Hopefully this patch will prevent similar screwups in the future.