Fix the manual to work with both texinfo-4 and texinfo-5.
Texinfo has been stable for the first 15 years of FFTW's history.
Then some genius, with too much time in his hands and on a mission to
deliver the world from the evil of the C language, decided to rewrite
makeinfo in Perl, the old C version of makeinfo being, as I said,
evil.  The official excuse for the rewrite was that now I can have my
manual in XML format, as if XML were a feature.

The result of this stroke of genius is that texinfo-5 has different
rules for macro expansion than texinfo-4 does, specifically regarding
whether or not spaces after a macro are ignored.  Texinfo-4 had weird
rules, but at least they were constant and internally more or less
consistent.  Texinfo-5 has different rules, and even worse the rules
in texinfo-5 are inconsistent between the TeX and HTML output
processors.  This situation makes it almost impossible for us to
produce a manual that works with both texinfo 4 and 5 in all modes
(TeX, info, and html).  The @noindent/@refill hack is my best shot at
patching this situation.