avoid multiple declarations of fftw_alignment_of()
fftw_alignment_of() was declared both in the API header file fftw3.h
and in the internal header file ifftw.h.  While there is nothing wrong
with this, it breaks the property that all exported symbols are
defined in the API directory.  E.g., I am not sure what happens on
windows without the proper DLLEXPORT nonsense.

To avoid any issues, rename the internal routine to
fftw_ialignment_of(), and define an API wrapper.