Added Power8 VSX SIMD support
Power8 is a descendent of Power7, but the switch to little endian
means the old altivec SIMD will not work due to shifts on load
and store, and the new VSX instructions are much improved.
This adds support for both single and double precision VSX SIMD,
using either gcc (tested with version 4.9) or IBM xlC (tested with
version 13.1.2, slower than gcc). Clang from llvm-3.7 is still too
buggy to compile VSX code correctly, but flags and detection has
been added so it will work with a correct clang.