Make 128/256 bit generic simd separate options
These will only be used on esoteric and/or new
architectures, which likely also miss cycle counters.
In this case the widest simd would be picked automatically
based on flops estimates, so to give the user more
control it is better to provide separate options
to enable/disable these two choices.