Update VSX SIMD to avoid inline assembly
Thanks to some help from Michael Gschwind of
IBM, this removes the remaining inline assembly
calls and replace the with vector functions. This
avoid interfering with the optimizer both on GCC
and XLC, and gets us another 3-10% of performance
when using VSX SIMD. Tested with GCC-4.9, XLC-13.1
in single and double on little-endian power 8.