Revert simd-avx.h changes from b606e3191
They didn't improve performance at all as far as I can tell,
and they ended up breaking the PGI compiler.

It is always tempting to use the fancy addsub instructions in FFTW to
do complex multiplications, but the reality is that FFTW is designed
to avoid complex multiplications in most cases (we started in the SSE
days), and thus they don't make any difference.  We are better off
using the minimal possible set of AVX instructions to minimize the
chance of triggering compiler bugs.

The same statement holds for _mm256_shuffle_pd() versus
_mm256_permute_pd(): in theory the latter is better, in practice
either one is rarely used.  However, SHUFFLE is older (since the SSE
days) and has a higher chance of working.