Assert that CURRENT-AGE=3
This is an attempt to prevent the 3.3.6 version screwup from occurring
again.

In any reasonable universe, libraries would have a version H and they
would specify a L such that the library is compatible with all
versions in [L..H].  Any sensible programmer would never change L, as
this breaks backward compatibility and screws users.  A new version
would increase H and be done.  Instead, libtool wants CURRENT=H and
AGE=H-L (a new version change two variables).  Furthermore, the name
of the library in the file system is a combination of L and H-L.  The
two changes of basis arent't even orthogonal.  Pure madness.

This change attempts to impose sanity by asserting that that the
implied L is 3, since we never intend to break backward compatibility
with fftw-3.3, which was version L=3.