Ditch --enable-debug-malloc and --enable-debug-alignment
We wrote DEBUG_MALLOC in 1997 to debug memory leaks.  Nowadays
DEBUG_MALLOC is just confusing.  Better tools are available, and
DEBUG_MALLOC is not thread-safe and it does not respect SIMD
alignment.  It confused at least one user.

In the gcc-2.SOMETHING days, gcc would allocate doubles on the stack
at 4-byte boundary (vs. 8) reducing performance by a factor of 3.
That's when we introduced --enable-debug-alignment, which is totally
obsolete by now.