Fix alignment requirements for avx512
The avx512 alignment requirement was set to 64 bytes, but this is
wrong.  Alignment requirements are a property of the platform (e.g.,
x86) and not of the instruction set (e.g., AVX).  Among other things,
this broke wisdom with avx512.

Good thing that avx512 was marked as experimental.  It's still
experimental, because I have no hardware to test it.

Thanks Tom Epperly for the bug report and debugging assistance.