Multi-width R5V, similar to arm-sve-clean.
This is a cleaned-up version of branch riscv-v from rdolbeau/fftw3.

No scalability, this is using multiple sets of masked operations to
support all width multiple of 128 bits.
This goes up to 16384 bits to support the EPAC VPU design.

This is now using the official V1.0 intirnsics, and was tested with gcc-14 and clang-18.
It can bes tested on qemu, or on hardware with V1.0 support.
It was thoroughly tested on a BananaPi F3 with a Spacemit K1 SoC, which supports V1.0 with 256-bits registers.