README.md: link to README
README.md refers the user to the text-only README, which contains - among other
things, build instructions. Since the two files have a very similar name, this
may lead to confusion for first-time users.

With this commit, a relative hyperlink pointing to README is added in README.md,
in order to make it explicit what file is being referenced.