First attempt at adding planner_nthreads (#205)
* First attempt at adding planner_nthreads

While you can currently set the maximum number of threads used by the planner,
there is no way to check that number. This simply adds a new function,
`planner_nthreads`, that provides this information, as was suggested by
@stevengj.

* Add basic testing of planner_nthreads

In an attempt to verify that planner_nthreads works as intended, I have added an
assertion in `fftw-bench.c` that `planner_nthreads` returns the nthreads just
passed as an argument to `plan_with_nthreads`.