Use -mavx2 instead of -march=core-avx2
It appears that -march=core-avx2 was necessary on some old version of
gcc-4.8, but setting -march is problematic for users who want to set
-march themselves.

It appears that whatever gcc bug caused us to use -march=core-avx2 in
2015 is long gone.  I have checked gcc-4.8.5 on centos-7 and
oraclelinux-7 (both 64-bit) and on 32-bit debian jessie and in all
cases -mavx2 is sufficient.

Therefore, given that the original cause for using -march is gone and
that using -march causes problems, I am removing the automatic
selection of -march=core-avx2 and I am using -mavx2 instead.  Users
for whom this is a problem can still override the default via

  ./configure --enable-avx2 AVX2_CFLAGS=...