cmake: Set SIMD flags on appropriate source files only.
Doing this matches the behavior of the configure script and makefiles.

Previously, enabling a set of SIMD instructions with the CMake build
would build the whole library with that flag enabled, allowing the
compiler to optimize by using that set of instructions anywhere. That
would then require that any SIMD instructions enabled at build time
would have to be present at run time. This change enables the intended
behavior of using SIMD instructions only when they are detected at run
time.